+++ packages/e2e-tests/tests/common/api-helpers/ownership-helper.ts
@@ -2,7 +2,10 @@ import { retry } from "@rarible/sdk/src/common/retry"
 import type { IRaribleSdk } from "@rarible/sdk/src/domain"
 import type { Ownership } from "@rarible/api-client/build/models"
 import type { Blockchain, GetOwnershipByIdResponse } from "@rarible/api-client"
-import type { BigNumber, ItemId } from "@rarible/types"
+import type { BigNumber, ContractAddress, ItemId } from "@rarible/types"
+import type { Ownerships } from "@rarible/api-client/build/models"
+import type {
+	GetOwnershipsByItemResponse } from "@rarible/api-client/build/apis/OwnershipControllerApi"
 
 
 export async function getOwnershipById(sdk: IRaribleSdk, blockchain: Blockchain, contractAddress: string,
@@ -28,7 +31,6 @@ export async function awaitForOwnershipValue(sdk: IRaribleSdk, itemId: ItemId,
 	expect(ownership).not.toBe(null)
 	if (value) {
 		expect(ownership.value).toBe(value)
-		console.log("value", ownership.value)
 	}
 	return ownership
 }
@@ -44,4 +46,28 @@ export async function getOwnershipByIdRaw(sdk: IRaribleSdk, itemId: ItemId,
 	expect(ownership.value as Ownership).not.toBe(null)
 	console.log(ownership.value)
 	return ownership
+}
+
+export async function getOwnershipsByItem(sdk: IRaribleSdk, contract: ContractAddress,
+	tokenId: BigNumber): Promise<Ownerships> {
+	const ownershipId = `${contract}:${tokenId}`
+	const ownerships = await retry(10, 2000, async () => {
+		return await sdk.apis.ownership.getOwnershipsByItem({
+			itemId: ownershipId,
+		})
+	})
+	expect(ownerships).not.toBe(null)
+	return ownerships
+}
+
+export async function getOwnershipsByItemRaw(sdk: IRaribleSdk, contract: ContractAddress,
+																						 tokenId: BigNumber): Promise<GetOwnershipsByItemResponse> {
+	const ownershipId = `${contract}:${tokenId}`
+	const ownerships = await retry(10, 2000, async () => {
+		return await sdk.apis.ownership.getOwnershipsByItemRaw({
+			itemId: ownershipId,
+		})
+	})
+	expect(ownerships).not.toBe(null)
+	return ownerships
 }