+++ packages/sdk/src/domain.ts
@@ -1,4 +1,5 @@
 import type * as ApiClient from "@rarible/api-client"
+import type { BlockchainGroup } from "@rarible/api-client"
 import type { Maybe } from "@rarible/types/build/maybe"
 import type { BlockchainWallet } from "@rarible/sdk-wallet"
 import type { IMint } from "./types/nft/mint/domain"
@@ -16,6 +17,7 @@ import type { IRestrictionSdk } from "./types/nft/restriction/domain"
 import type { IPreprocessMeta } from "./types/nft/mint/preprocess-meta"
 import type { Middleware } from "./common/middleware/middleware"
 import type { RaribleSdkEnvironment } from "./config/domain"
+import type { ISolanaSdkConfig } from "./sdk-blockchains/solana/domain"
 
 export enum LogsLevel {
 	DISABLED = 0,
@@ -32,6 +34,9 @@ export interface ISdkContext {
 export interface IRaribleSdkConfig {
 	apiClientParams?: ApiClient.ConfigurationParameters
 	logs?: LogsLevel
+	blockchain?: {
+		[BlockchainGroup.SOLANA]?: ISolanaSdkConfig
+	}
 	middlewares?: Middleware[]
 }